+++ xmldiff/formatting.py
@@ -13,6 +13,7 @@ DIFF_PREFIX = "diff"
 
 INSERT_NAME = "{%s}insert" % DIFF_NS
 DELETE_NAME = "{%s}delete" % DIFF_NS
+REPLACE_NAME = "{%s}replace" % DIFF_NS
 RENAME_NAME = "{%s}rename" % DIFF_NS
 
 # Flags for whitespace handling in the text aware formatters:
@@ -104,9 +105,14 @@ class PlaceholderMaker:
         delete_close = self.get_placeholder(delete_elem, T_CLOSE, None)
         delete_open = self.get_placeholder(delete_elem, T_OPEN, delete_close)
 
+        replace_elem = etree.Element(REPLACE_NAME)
+        replace_close = self.get_placeholder(replace_elem, T_CLOSE, None)
+        replace_open = self.get_placeholder(replace_elem, T_OPEN, replace_close)
+
         self.diff_tags = {
             "insert": (insert_open, insert_close),
             "delete": (delete_open, delete_close),
+            "replace": (replace_open, replace_close),
         }
 
     def get_placeholder(self, element, ttype, close_ph):
@@ -228,7 +234,7 @@ class PlaceholderMaker:
     def undo_tree(self, tree):
         self.undo_element(tree)
 
-    def mark_diff(self, ph, action):
+    def mark_diff(self, ph, action, attributes=None):
         entry = self.placeholder2tag[ph]
         if entry.ttype == T_CLOSE:
             # Close tag, nothing to mark
@@ -242,12 +248,22 @@ class PlaceholderMaker:
             # Formatting element, add a diff attribute
             action += "-formatting"
         elem.attrib[f"{{{DIFF_NS}}}{action}"] = ""
+        if attributes is not None:
+            for attrib, value in attributes.items():
+                elem.attrib[attrib] = value
 
         # And make a new placeholder for this new entry:
         return self.get_placeholder(elem, entry.ttype, entry.close_ph)
 
-    def wrap_diff(self, text, action):
+    def wrap_diff(self, text, action, attributes=None):
         open_ph, close_ph = self.diff_tags[action]
+        if attributes is not None and len(attributes) > 0:
+            entry = self.placeholder2tag[open_ph]
+            elem = entry.element
+            elem = deepcopy(elem)
+            for attrib, value in attributes.items():
+                elem.attrib[attrib] = value
+            open_ph = self.get_placeholder(elem, entry.ttype, entry.close_ph)
         return open_ph + text + close_ph
 
 
@@ -295,16 +311,25 @@ class XMLFormatter(BaseFormatter):
     WS_TEXT normalizes only inside text_tags, WS_TAGS will remove ignorable
     whitespace between tags, WS_BOTH do both, and WS_NONE will preserve
     all whitespace.
+
+    The ``use_replace`` flag decides, if a replace tag (with the old text
+    as an attribute) should be used instead of one delete and one insert tag.
     """
 
     def __init__(
-        self, normalize=WS_NONE, pretty_print=True, text_tags=(), formatting_tags=()
+        self,
+        normalize=WS_NONE,
+        pretty_print=True,
+        text_tags=(),
+        formatting_tags=(),
+        use_replace=False,
     ):
         # Mapping from placeholders -> structural content and vice versa.
         self.normalize = normalize
         self.pretty_print = pretty_print
         self.text_tags = text_tags
         self.formatting_tags = formatting_tags
+        self.use_replace = use_replace
         self.placeholderer = PlaceholderMaker(
             text_tags=text_tags, formatting_tags=formatting_tags
         )
@@ -570,6 +595,36 @@ class XMLFormatter(BaseFormatter):
                         new_diff.append((op, seg))
         return new_diff
 
+    def _join_delete_insert(self, diffs):
+        new_diffs = []
+        skip_next = False
+        for i in range(len(diffs) - 1):
+            if skip_next:
+                skip_next = False
+                continue
+            op, text = diffs[i]
+            next_op, next_text = diffs[i + 1]
+            # insert, then delete
+            if (
+                op == diff_match_patch.DIFF_INSERT
+                and next_op == diff_match_patch.DIFF_DELETE
+            ):
+                new_diffs.append((diff_match_patch.DIFF_REPLACE, text, next_text))
+                skip_next = True  # also skip upcoming delete
+            # delete, then insert
+            elif (
+                next_op == diff_match_patch.DIFF_INSERT
+                and op == diff_match_patch.DIFF_DELETE
+            ):
+                new_diffs.append((diff_match_patch.DIFF_REPLACE, next_text, text))
+                skip_next = True  # also skip upcoming insert
+            else:
+                new_diffs.append(diffs[i])
+        # append last diff, if it shouldn't be skipped
+        if not skip_next:
+            new_diffs.append(diffs[-1])
+        return new_diffs
+
     def _make_diff_tags(self, left_value, right_value, node, target=None):
         if bool(self.normalize & WS_TEXT):
             left_value = utils.cleanup_whitespace(left_value or "").strip()
@@ -578,36 +633,46 @@ class XMLFormatter(BaseFormatter):
         text_diff = diff_match_patch()
         diff = text_diff.diff_main(left_value or "", right_value or "")
         text_diff.diff_cleanupSemantic(diff)
-
         diff = self._realign_placeholders(diff)
 
+        if self.use_replace:
+            diff = self._join_delete_insert(diff)
         cur_child = None
         if target is None:
             target = node
         else:
             cur_child = node
 
-        for op, text in diff:
-            if op == 0:
+        for d in diff:
+            op = d[0]
+            text = d[1]
+            if op == diff_match_patch.DIFF_REPLACE:
+                old_text = d[2]
+
+            if op == diff_match_patch.DIFF_EQUAL:
                 if cur_child is None:
                     node.text = (node.text or "") + text
                 else:
                     cur_child.tail = (cur_child.tail or "") + text
                 continue
 
-            if op == -1:
+            attributes = {}
+            if op == diff_match_patch.DIFF_DELETE:
                 action = "delete"
-            elif op == 1:
+            elif op == diff_match_patch.DIFF_INSERT:
                 action = "insert"
+            elif op == diff_match_patch.DIFF_REPLACE:
+                action = "replace"
+                attributes["old-text"] = old_text
 
             if self.placeholderer.is_placeholder(text):
-                ph = self.placeholderer.mark_diff(text, action)
+                ph = self.placeholderer.mark_diff(text, action, attributes)
 
                 if cur_child is None:
                     node.text = (node.text or "") + ph
 
             else:
-                new_text = self.placeholderer.wrap_diff(text, action)
+                new_text = self.placeholderer.wrap_diff(text, action, attributes)
 
                 if cur_child is None:
                     node.text = (node.text or "") + new_text